+++ examples/Get-GPPPassword.py
@@ -257,19 +257,6 @@ def parse_target(args):
 
     return domain, username, password, address, lmhash, nthash
 
-
-def init_logger(args):
-    # Init the example's logger theme and debug level
-    logger.init(args.ts)
-    if args.debug is True:
-        logging.getLogger().setLevel(logging.DEBUG)
-        # Print the Library's installation path
-        logging.debug(version.getInstallationPath())
-    else:
-        logging.getLogger().setLevel(logging.INFO)
-        logging.getLogger("impacket.smbserver").setLevel(logging.ERROR)
-
-
 def init_smb_session(args, domain, username, password, address, lmhash, nthash):
     smbClient = SMBConnection(address, args.target_ip, sess_port=int(args.port))
     dialect = smbClient.getDialect()
@@ -295,7 +282,7 @@ def init_smb_session(args, domain, username, password, address, lmhash, nthash):
 if __name__ == '__main__':
     print(version.BANNER)
     args = parse_args()
-    init_logger(args)
+    logger.init(args.ts, args.debug)
 
     if args.target.upper() == "LOCAL":
         if args.xmlfile is not None:
