+++ examples/getPac.py
@@ -40,7 +40,7 @@ from pyasn1.type.univ import noValue
 from impacket import version
 from impacket.dcerpc.v5.rpcrt import TypeSerialization1
 from impacket.examples import logger
-from impacket.examples.utils import parse_credentials
+from impacket.examples.utils import parse_identity
 from impacket.krb5 import constants
 from impacket.krb5.asn1 import AP_REQ, AS_REP, TGS_REQ, Authenticator, TGS_REP, seq_set, seq_set_iter, PA_FOR_USER_ENC, \
     EncTicketPart, AD_IF_RELEVANT, Ticket as TicketAsn1
@@ -292,7 +292,6 @@ class S4U2SELF:
 
 # Process command-line arguments.
 if __name__ == '__main__':
-    logger.init()
     print(version.BANNER)
 
     parser = argparse.ArgumentParser()
@@ -301,6 +300,7 @@ if __name__ == '__main__':
                                                        'for grabbing targetUser\'s PAC')
     parser.add_argument('-targetUser', action='store', required=True, help='the target user to retrieve the PAC of')
     parser.add_argument('-debug', action='store_true', help='Turn DEBUG output ON')
+    parser.add_argument('-ts', action='store_true', help='Adds timestamp to every logging output')
 
     group = parser.add_argument_group('authentication')
 
@@ -310,22 +310,9 @@ if __name__ == '__main__':
         sys.exit(1)
 
     options = parser.parse_args()
+    logger.init(options.ts, options.debug)
 
-    domain, username, password = parse_credentials(options.credentials)
-
-    if domain is None:
-        domain = ''
-
-    if password == '' and username != '' and options.hashes is None:
-        from getpass import getpass
-        password = getpass("Password:")
-
-    if options.debug is True:
-        logging.getLogger().setLevel(logging.DEBUG)
-        # Print the Library's installation path
-        logging.debug(version.getInstallationPath())
-    else:
-        logging.getLogger().setLevel(logging.INFO)
+    domain, username, password, _, _, _ = parse_identity(options.credentials, options.hashes)
 
     try:
         dumper = S4U2SELF(options.targetUser, username, password, domain, options.hashes)
