+++ examples/kintercept.py
@@ -38,6 +38,8 @@ from impacket.krb5.crypto import Cksumtype
 from impacket.krb5.asn1 import TGS_REQ, TGS_REP, seq_set, PA_FOR_USER_ENC
 from impacket.krb5.types import Principal
 
+from impacket.examples import logger
+
 
 MAX_READ_SIZE = 16000
 MAX_BUFF_SIZE = 32000
@@ -267,6 +269,8 @@ def parse_args():
     parser.add_argument('--listen-addr', default='', help='Address to listen on')
     parser.add_argument('--request-handler', default='', metavar='HANDLER:ARG', help='Example: s4u2else:user')
     parser.add_argument('--reply-handler', default='', metavar='HANDLER:ARG', help='Example: tgs-rep-user:user')
+    parser.add_argument('-ts', action='store_true', help='Adds timestamp to every logging output')
+    parser.add_argument('-debug', action='store_true', help='Turn DEBUG output ON')
     return vars(parser.parse_args())
 
 
@@ -275,6 +279,8 @@ if __name__ == '__main__':
     print(version.BANNER)
 
     args = parse_args()
+
+    logger.init(args.ts, args.debug)
 
     req_factory = rep_factory = InterceptConnFactory()
     if args['request_handler']:
